refactor(notes): types
