+++ src/modules/notes/Note/HeaderNote/service/HeaderNote.service.ts
@@ -1,10 +1,11 @@
-import { IFetchResponseDefault } from "shared/common/types/Fetch";
 import api from "shared/services/api";
+import { IFetchPostCreateNoteResponse, IFetchPostCreateNoteService } from "../types";
 
-const fetchPostCreateNoteService = async (data: any) => {
+const fetchPostCreateNoteService = async (data: IFetchPostCreateNoteService) => {
 	const { board_id, payload } = data;
-	const response = await api.post<IFetchResponseDefault>(`notes/create/board_id=${board_id}`, payload);
-	return response;
+	const response = await api.post<IFetchPostCreateNoteResponse>(`notes/create/board_id=${board_id}`, payload);
+	
+	return response.data;
 };
 
 export { fetchPostCreateNoteService };
