--- src/modules/notes/Note/HeaderNote/types/HeaderNote.component.ts
@@ -1,9 +1,0 @@
-import { IFetchGetSingleBoard } from "shared/common/types/Fetch";
-
-export interface IHeaderNoteView {
-    data: IFetchGetSingleBoard | undefined;
-    handleClickCreateNote: () => void;
-    openDialogDeleteAllNotes: () => void;
-    isMediumScreen: boolean;
-    countNotes: number;
-}
