+++ src/modules/notes/Note/HeaderNote/types/index.ts
@@ -1,0 +1,2 @@
+export * as default from "./HeaderNote.types";
+export * from "./HeaderNote.types";
